Major error in modulized_time_delay_proto.m and sigma->sqrt(4*D*dt)
On line 93 and 102 there the end function was misplaced. If you check the version on the major branch 2020.10.2-second-modulization you can spot the error. This error will let the positions x and y become 0 after k=1. Also the random number generator in modulized_time_delay_proto.m was changed of its sigma=sqrt(2*D*dt) to sigma=sqrt(4*D*dt) to match the values in Tong. Comparing the graphs generated by main.m the x_rms in the simulation seems to be a larger than the theoretical curve x_rms=sqrt(6*D*t)