Corrections for cases with movie_create='off'
Main.m: If movie_create='off', then the function save_movie on line 54 cannot function because the MovieVector is empty. We have fixed this issue for the branch 2020.10.6-test-dt-deterministic, and decided to update this one, but haven't tested for this branch yet.

make_movie_plots.m: for movie_create='off' the MovieVector will be returned with an empty set.