Issue solved no overla[
I implemented a relaxation for second_stage_delayed_int.m  in the end of every update of k. This has prevented the overlapping phenomenon back from last week. The coefficient on line 105~108 is chosen to be 1; However, 0.5 is the one that makes the particles touch each other, but if we choose 0.5 the relaxation is then too slow.